requirements.txt
azure-functions
azure-storage-queue
pymssql



